% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdm_demo.R
\name{vdm_demo}
\alias{vdm_demo}
\title{calculate the Value Difference Metric}
\usage{
vdm_demo(x, y, norm = 2)
}
\arguments{
\item{x, }{a string or factor}

\item{y, }{the corresponding class label}

\item{norm, }{the norm to use}
}
\value{
the vdm value between pairs of the levels of x
}
\description{
calculate the Value Difference Metric
}
\details{
VDM Calculation Demonstration

More details about VDM can be found at:
https://en.wikipedia.org/wiki/Value_difference_metric

Part of the code is adapted from the following source:
https://github.com/esmondhkchu/rvdm/tree/master
}
\examples{
x <- xigua3.0$根蒂
y <- xigua3.0$好瓜
vdm_demo(x, y)

}
